PATCH: [API: user#create/update] Expose preferences attribute
For some reason this is called "settings" in the web API
but "preferences" in the object model.

SECURITY CONSIDERATIONS:
Are we checking that the client is authorized to update
the user?  It doesn't appear so, but the upstream API
allows changing far more sensitive properties than just
the user’s mere preferences.