PATCH: Include room name in room-opened event
The event is handled by sapi and shows the room name on the page
hosting the iframe.

Question: Why doesn't the room know its own name?

Implementation:
I'd prefer to change the order of the code:
- clone the room
- add the name
- and then omit the usernames in the fireGlobalEvent call,
  as the original code did.
But I don't know the lowdash library well-enough to know how to clone.
TODO: learn how and refactor.