PATCH: [OBSOLETE?]  Store current room name/type in Session
As near as I can tell, these lines are vestigial.
They were originally added as part of
  commit 4bbdd6885291f7264726fe0c09c57843b5238302 (in 0.73.1-sa.1)
  "Send mone event directly to SA. Mone events for sent-message, room-opened, etc."

but the lines that use them -- added in that same commit to
  packages/rocketchat-lib/client/methods/sendMessage.js
  packages/rocketchat-ui/client/views/app/room.js
have since been dropped (but I can't find where).

DEPENDENCY:
This patch depends on "PATCH: Include room name in room-opened event"